requirements.txt
fastapi
uvicorn
groq
pydantic
python-dotenv
requests
transformers
torch
scikit-learn
numpy
pandas
sentence-transformers